requirements.txt
asgiref==3.9.1
channels==4.3.1
channels_redis==4.3.0
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
msgpack==1.1.1
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
redis==6.4.0
sqlparse==0.5.3
tzdata==2025.2
